infra/agents/plan-generator/pyproject.toml
[project]
name = "plan-generator"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "strands-agents>=1.0.0",
    "fitness-engine",
    "fitness-contracts",
    "boto3>=1.34.0",
    "pydantic>=2.7.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[tool.uv.sources]
fitness-engine = { path = "../../../packages/fitness-engine", editable = true }
fitness-contracts = { path = "../../../packages/contracts-py", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src/plan_generator"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "moto[dynamodb]>=5.0"]
